Generalize cloned handler placement
Use fgNewBBinRegion to place the first clone handler block, so
that we get the right placement for different EH implementations.

Without callfinally thunks, callfinallies might appear in handler
regions -- eg in the catch of a try-catch inside a try-finally.
So relax the not in handler assert for those cases.

Verify the EH table after both finally cloning and empty finally
removal.

Screen for callfinallies more carefully in empty finally removal.