More x86 fixes: clear finalStep flag, get right try region.
Previous commit set things up to put the cloned finally in the
right try region, but we were still using the region of the callfinally
which is incorrect for x86. Use the region of the finally instead.

We may see BBF_KEEP_BBJ_ALWAYS flag bits on the continuation blocks
for finallies (the "finalStep" blocks). If we remove the finally, then
clear the flag bits.

x86 now passing CoreCLR tests...