Try exit path selection and clone placement
Modify how we look for the try exit path: instead of scanning the
clonefinally range, walk the try region backwards to find the last
invocation of the finally from the try. Then move the corresponding
callfinally to the start of the callfinally range, so that we're more
likely to be able to fall from the try into the cloned finally.

Also reworked some local names, removed an unused local, and added
a few more asserts.