Convert small fixed-sized locallocs to local vars
During importation, if the argument to localloc is a small constant
between 0 and 32, create a new TYP_BLK local for the allocation, and
do not set `compLocallocUsed`.

Max size was chosen based on the cases seen in CoreCLR core library
and frameworks.

This optimization fires in various IL stubs. Haven't seen too many
examples from user-written code.