Jit: unsafe devirtualization
Prototype where the jit will aggressively devirtualize based on what is
known about the loaded classes at the time a method is jitted.

Breaks crossgen assumption about classes being "pre-restored" so one must
delete system.private.corlib.ni.dll before running anything. Even with that
we are bitten during app launch in system code and so devirtualization is
disabled in the system assembly for now.

Might be able to get farther with finer-grained tracking, eg when jitting method
B:A, if we devirt a call to C:F based on class state of C, we only need to invalidate
B:A if some superclass X of C overrides F and an X object is actually created.

Right now we'd bail as soon as we saw any subclass of C being loaded since the devirt
flips a bit on C indicating we assumed no subclasses.

We also don't track the codegen dependence, so when the assumption is violated we
don't know which methods are suspect. We could capture this somehow and see if we
could toss out method bodies and rejit. It would work provided we don't need OSR.