JIT: optimize type tests using class info
Proof of concept that we can optimize away type tests using the
more generalized class info tracking that was introduced for
devirtualization.

If the class of a tree node `obj` is known exactly (via `gtGetClassHandle`)
then  `obj.GetType() == typeof(...)` can be optimized at jit time to
either true or false. There may still need to be an upstream null check
on `obj`.

If `obj`'s type is not known exactly, the value returned by `gtGetClassHandle`
may be usable to accelerate the runtime type check since it may be a better
approximation of the true type (nyi, maybe not worth the trouble).

Also includes changes to recognize when we have exact types in more cases,
eg any final non-variant non-array ref type is exact, as are arrays of such
types. In particular `string[]` is an exact type.