Update CheckProjects to not fail for non-dev CORE_ROOT
CheckProjects expects to be able to find the jit test project files
relative to the value of the `CORE_ROOT` environment variable. This
fails when we have test setups like cross-tests or helix where the
test binary tree has been copied after building.

Tolerate these cases by returning a success code if the `CORE_ROOT`
path is not pointing in the expected place.