Revise implementation of Enum.HasFlag
Move all the argument validation for Enum.HasFlag over to the native code
side, in anticpation of enabling the jit to optimize these calls.

Add a test to make sure a properly formatted exception message is created
for argument type mismatches.

Also add a note to Enum.cs about the half-baked support for float/double
and pointer sized enums.