Work on type equality testing in the jit interface
Add two new interface methods, one for testing for type equality and
the other for type casting.

These return Must/MustNot/Maybe results depending on whether or not
the equality or cast can be resolved at jit time.

Implmement a first cut at the body of the equality check. The casting
check is still stubbed out and always returns Maybe.