Fix a couple issues seen in initial testing
`gtOptimizeEnumHasFlag` returns null on failure, not the original tree.
So check for this.

A `refanytype` creates runtime handle to type helper call with an indir
off of the refany, this isn't marked as nonfaulting. Perhaps it should
be, but even if so we wouldn't optimize further. So check for faulting
indirs instead of asserting.

Also I am pretty sure the helper properties for the handle to type helpers
could be improved: the MAYBENULL helper isn't even in the table and the
normal will never return null. But will defer changing this for now.