JIT: handle cases where importer folding can lead to mistyping
When the jit is optimizing and folds a branch in the importer, some
statically reachable blocks may become unreachable, and the jit may never
import them.

This can lead to optimization-induced differences in types on the
evaluation stack in reachable blocks and hence observable differences
in the behavior of jit-generated code.

For instance given an ABCD diamond, if A's branch is foldable, B leaves
an int32 on the stack, and C leaves a native int, then the stack on entry
to D could be typed as int32 when optimizing but native int when not
optimizing.

The jit can detect when this sort of mis-typing is possible by noting
if any branches are folded, and noting if any block leaves an int32
(in 64 bit mode) or float value on the stack.

When this situation is detected, at the end of normal importation, the
jit disables branch folding and reimports all the code in the method.
Because folding is disabled the jit will see all the type contributions
in the second importation.

Empirically these cases of potential mis-typing are rare. Only 0.5% of
methods in the jit-diff fx set trigger the reimportation.