JIT: don't use VirtualAlloc for jit allocations for CoreCLR
Jit throughput is impacted by use of VirtualAlloc for page allocations
This can be avoided by simply using a normal heap for allocations.

There is potentially some added risk here as jit allocations commingle
with other heap allocations, for apps that might be misusing the heap.

Keep desktop behavior as is for now.