JIT: prototype support for guarded devirtualization
Gives the jit the ability to transform a virtual or interface call into a class
test that guards a devirtualized call, with a fallback path that invokes the
unaltered virtual or interface call.

See the included document for some background and motivation.

Currently uses the jit's "best known" class as the class to guess for in the
guard test.

The devirtualized call can subsequently be inlined.

Still a long list of TODOs, including but not limited to:
* better heuristics for when to transform calls
* better heuristics for the class to test for
* proper handling of all the struct return cases
* cleaner interaction with late devirtualization
* possible undo if the subsequent inline attempt fails
* add the ability for one class test to cover several calls