JIT: look for more cases where covariant store check can be optimized
Re #13574

Once we have SSA and Value Numbers, we can more generally scan for cases where
an array element is being set to a value taken from the same array.

In such cases there is no need for a covariant store check (a write barrier
and bounds check may still be needed).

This allows us to optimize a number of element swap patterns we currently miss
via the simplistic analysis possible in the importer.

This doesn't really fit in assertionprop as it does not rely on assertion
information, but I don't see it meriting a full pass of its own.

There are more cases like this we could optimize if we knew the exact type of
the array or the exact type of the element being stored in the array.