Block tail recursion to loop for OSR
Can't handle this (yet) because the entry blocks of the method are
not imported, so there is no place one can branch to.

Could fix with whole-method OSR importation, or perhaps lighting
up the entry blocks when this case is detected. Note early on we are
not sure if we will transform or not; wonder if those lit-up entry
blocks would then be removed...